Move SaltStackRunResults to results package
