+++ src/main/java/com/suse/saltstack/netapi/results/SaltStackRunResults.java
@@ -1,4 +1,4 @@
-package com.suse.saltstack.netapi.client;
+package com.suse.saltstack.netapi.results;
 
 import java.util.List;
 import java.util.Map;
